+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/commands/SlimefunCommand.java
@@ -1,7 +1,9 @@
 package io.github.thebusybiscuit.slimefun4.core.commands;
 
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.bukkit.ChatColor;
@@ -26,6 +28,7 @@ public class SlimefunCommand implements CommandExecutor, Listener {
 
     private final SlimefunPlugin plugin;
     private final List<SubCommand> commands = new LinkedList<>();
+    private final Map<SubCommand, Integer> commandUsage = new HashMap<>();
 
     public SlimefunCommand(SlimefunPlugin plugin) {
         plugin.getServer().getPluginManager().registerEvents(this, plugin);
@@ -39,6 +42,10 @@ public class SlimefunCommand implements CommandExecutor, Listener {
 
     public SlimefunPlugin getPlugin() {
         return plugin;
+    }
+
+    public Map<SubCommand, Integer> getCommandUsage() {
+        return commandUsage;
     }
 
     @Override
@@ -78,6 +85,7 @@ public class SlimefunCommand implements CommandExecutor, Listener {
             else {
                 for (SubCommand command : commands) {
                     if (args[0].equalsIgnoreCase(command.getName())) {
+                        commandUsage.merge(command, 1, Integer::sum);
                         command.onExecute(sender, args);
                         return true;
                     }