+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/metrics/ResearchesEnabledChart.java
@@ -1,0 +1,16 @@
+package io.github.thebusybiscuit.slimefun4.core.services.metrics;
+
+import org.bstats.bukkit.Metrics.SimplePie;
+
+import me.mrCookieSlime.Slimefun.SlimefunPlugin;
+
+class ResearchesEnabledChart extends SimplePie {
+
+    public ResearchesEnabledChart() {
+        super("servers_with_researches_enabled", () -> {
+            boolean enabled = SlimefunPlugin.getSettings().researchesEnabled;
+            return enabled ? "enabled" : "disabled";
+        });
+    }
+
+}